Fix running tests and make the environment more "nodocker-friendly"

+++ hack/buildenv.sh
@@ -11,13 +11,17 @@ if [[ ! -e vpp/Makefile ]]; then
   exit 1
 fi
 
-. hack/build-image-name.sh
-
-docker run -it --rm --name vpp-build --shm-size 1024m \
-       -v $PWD/vpp:/src/vpp:delegated \
-       -v $PWD/upf:/src/upf:delegated \
-       -v $PWD/vpp-out:/vpp-out \
-       -e LC_ALL=C.UTF-8 -e LANG=C.UTF-8 \
-       -w /src/vpp \
-       "${build_image}" \
-       "$@"
+if [[ ${UPF_NO_DOCKER_BUILDENV:-} ]]; then
+  cd vpp
+  exec "$@"
+else
+  . hack/build-image-name.sh
+  docker run -it --rm --name vpp-build --shm-size 1024m \
+         -v $PWD:/src:delegated \
+         -v $PWD/vpp-out:/vpp-out \
+         -e LC_ALL=C.UTF-8 -e LANG=C.UTF-8 \
+         -e UPF_NO_DOCKER_BUILDENV=1 \
+         -w /src/vpp \
+         "${build_image}" \
+         "$@"
+fi
